+++ CHANGELOG.md
@@ -8,6 +8,13 @@ Release 5.7.0:
  - Holder:
    - Replace `keyPair` with `keyMaterial`
 
+Release 5.6.1:
+ - Expose details for `ConstraintFieldsEvaluationException`
+ - Token status:
+   - Errors in status list lookup lead to a `null` token status, not to an error as before, i.e. `TokenStatusEvaluationException` is never thrown
+ - Remote Qualified Electronic Signatures:
+   - In `RqesOpenId4VpHolder` fix validation of signing credentials
+
 Release 5.6.0:
  - Remote Qualified Electronic Signatures:
    - Fix erroneous `InputDescriptor` encoding in `PresentationDefinition` when more specific type  was known (i.e. `DidInputDescriptor`/`QesInputDescriptor`) via contexutal serialziation